feat!: clean up deprecated Databases and DataTypes, remove migrations older for versions older than 12 months
