+++ meteor/server/migration/0_1_0.ts
@@ -1,14 +1,422 @@
 import { addMigrationSteps } from './databaseMigration'
 import { logger } from '../logging'
-import { getRandomId, protectString } from '../../lib/lib'
+import { getRandomId, protectString, generateTranslation as t, getHash } from '../../lib/lib'
 import { wrapDefaultObject } from '@sofie-automation/corelib/dist/settings/objectWithOverrides'
 import { ShowStyleVariantId } from '@sofie-automation/corelib/dist/dataModel/Ids'
-import { ShowStyleBases, ShowStyleVariants, Studios } from '../collections'
+import { ShowStyleBases, ShowStyleVariants, Studios, TriggeredActions } from '../collections'
+import {
+	IBlueprintTriggeredActions,
+	ClientActions,
+	TriggerType,
+	PlayoutActions,
+} from '@sofie-automation/blueprints-integration'
 
 /**
  * This file contains system specific migration steps.
  * These files are combined with / overridden by migration steps defined in the blueprints.
  */
+
+let j = 0
+
+const DEFAULT_CORE_TRIGGERS: IBlueprintTriggeredActions[] = [
+	{
+		_id: 'core_toggleShelf',
+		actions: {
+			'0': {
+				action: ClientActions.shelf,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				state: 'toggle',
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Tab',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Toggle Shelf'),
+	},
+	{
+		_id: 'core_activateRundownPlaylist',
+		actions: {
+			'0': {
+				action: PlayoutActions.activateRundownPlaylist,
+				rehearsal: false,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Backquote',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Activate (On-Air)'),
+	},
+	{
+		_id: 'core_activateRundownPlaylist_rehearsal',
+		actions: {
+			'0': {
+				action: PlayoutActions.activateRundownPlaylist,
+				rehearsal: true,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Control+Backquote',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Activate (Rehearsal)'),
+	},
+	{
+		_id: 'core_deactivateRundownPlaylist',
+		actions: {
+			'0': {
+				action: PlayoutActions.deactivateRundownPlaylist,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Control+Shift+Backquote',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Deactivate'),
+	},
+	{
+		_id: 'core_take',
+		actions: {
+			'0': {
+				action: PlayoutActions.take,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'NumpadEnter',
+				up: true,
+			},
+			'1': {
+				type: TriggerType.hotkey,
+				keys: 'F12',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Take'),
+	},
+	{
+		_id: 'core_hold',
+		actions: {
+			'0': {
+				action: PlayoutActions.hold,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'KeyH',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Hold'),
+	},
+	{
+		_id: 'core_hold_undo',
+		actions: {
+			'0': {
+				action: PlayoutActions.hold,
+				undo: true,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Shift+KeyH',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Undo Hold'),
+	},
+	{
+		_id: 'core_reset_rundown_playlist',
+		actions: {
+			'0': {
+				action: PlayoutActions.resetRundownPlaylist,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Control+Shift+F12',
+				up: true,
+			},
+			'1': {
+				type: TriggerType.hotkey,
+				keys: 'Control+Shift+AnyEnter',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Reset Rundown'),
+	},
+	{
+		_id: 'core_disable_next_piece',
+		actions: {
+			'0': {
+				action: PlayoutActions.disableNextPiece,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'KeyG',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Disable the next element'),
+	},
+	{
+		_id: 'core_disable_next_piece_undo',
+		actions: {
+			'0': {
+				action: PlayoutActions.disableNextPiece,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				undo: true,
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Shift+KeyG',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Undo Disable the next element'),
+	},
+	{
+		_id: 'core_create_snapshot_for_debug',
+		actions: {
+			'0': {
+				action: PlayoutActions.createSnapshotForDebug,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Backspace',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Store Snapshot'),
+	},
+	{
+		_id: 'core_move_next_part',
+		actions: {
+			'0': {
+				action: PlayoutActions.moveNext,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				parts: 1,
+				segments: 0,
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'F9',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Move Next forwards'),
+	},
+	{
+		_id: 'core_move_next_segment',
+		actions: {
+			'0': {
+				action: PlayoutActions.moveNext,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				parts: 0,
+				segments: 1,
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'F10',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Move Next to the following segment'),
+	},
+	{
+		_id: 'core_move_previous_part',
+		actions: {
+			'0': {
+				action: PlayoutActions.moveNext,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				parts: -1,
+				segments: 0,
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Shift+F9',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Move Next backwards'),
+	},
+	{
+		_id: 'core_move_previous_segment',
+		actions: {
+			'0': {
+				action: PlayoutActions.moveNext,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+				parts: 0,
+				segments: -1,
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Shift+F10',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Move Next to the previous segment'),
+	},
+	{
+		_id: 'core_go_to_onAir_line',
+		actions: {
+			'0': {
+				action: ClientActions.goToOnAirLine,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Control+Home',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Go to On Air line'),
+	},
+	{
+		_id: 'core_rewind_segments',
+		actions: {
+			'0': {
+				action: ClientActions.rewindSegments,
+				filterChain: [
+					{
+						object: 'view',
+					},
+				],
+			},
+		},
+		triggers: {
+			'0': {
+				type: TriggerType.hotkey,
+				keys: 'Shift+Home',
+				up: true,
+			},
+		},
+		_rank: ++j * 1000,
+		name: t('Rewind segments to start'),
+	},
+]
 
 // 0.1.0: These are the "base" migration steps, setting up a default system
 export const addSteps = addMigrationSteps('0.1.0', [
@@ -116,6 +524,34 @@ export const addSteps = addMigrationSteps('0.1.0', [
 					_rank: 0,
 				})
 			}
+		},
+	},
+	{
+		id: 'TriggeredActions.core',
+		canBeRunAutomatically: true,
+		validate: () => {
+			const coreTriggeredActionsCount = TriggeredActions.find({
+				showStyleBaseId: null,
+			}).count()
+
+			if (coreTriggeredActionsCount === 0) {
+				return `No system-wide triggered actions set up.`
+			}
+
+			return false
+		},
+		migrate: () => {
+			DEFAULT_CORE_TRIGGERS.forEach((triggeredAction) => {
+				TriggeredActions.insert({
+					_id: protectString(getHash(triggeredAction._id)),
+					_rank: triggeredAction._rank,
+					name: triggeredAction.name,
+					blueprintUniqueId: null,
+					showStyleBaseId: null,
+					actionsWithOverrides: wrapDefaultObject(triggeredAction.actions),
+					triggersWithOverrides: wrapDefaultObject(triggeredAction.triggers),
+				})
+			})
 		},
 	},
 ])
