+++ meteor/server/migration/migrations.ts
@@ -3,24 +3,6 @@
 import { addSteps as addSteps0_1_0 } from './0_1_0'
 addSteps0_1_0()
 
-import { addSteps as addSteps1_19_0 } from './1_19_0'
-addSteps1_19_0()
-
-import { addSteps as addSteps1_32_0 } from './1_32_0'
-addSteps1_32_0()
-
-import { addSteps as addSteps1_35_0 } from './1_35_0'
-addSteps1_35_0()
-
-import { addSteps as addSteps1_36_0 } from './1_36_0'
-addSteps1_36_0()
-
-import { addSteps as addSteps1_37_0 } from './1_37_0'
-addSteps1_37_0()
-
-import { addSteps as addSteps1_38_0 } from './1_38_0'
-addSteps1_38_0()
-
 import { addSteps as addSteps1_39_0 } from './1_39_0'
 addSteps1_39_0()
